Adicionado configuracao de cors, ajuste de migrations e do mapper PersonMapper

+++ src/main/resources/db/migration/V3__Create_Table_Books.sql
@@ -1,6 +1,6 @@
 ﻿CREATE TABLE IF NOT EXISTS books
 (
-    id bigint NOT NULL,
+    id serial NOT NULL,
     author varchar,
     launch_date timestamp NOT NULL,
     price decimal(65,2) NOT NULL,
